discord/requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
Pillow>=10.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
